mved basic tests to their own file
